test: Move ensure_directory's tests to dedicated file
Instead of squashing them into the mount_unmount helpers, the dedicated
tests for ensure_directory are moved to a dedicated file. While at it, a
new test was written which ensured correct functionality even for parent
directories owned by root.
